backend/requirements.txt
asn1crypto==0.24.0
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
botocore==1.17.51
certifi==2018.1.18
chalice==1.18.1
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
enum-compat==0.0.3
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lightdm-gtk-greeter-settings==1.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
netifaces==0.10.4
oauth==1.0.1
peewee==3.13.3
pexpect==4.2.1
Pillow==5.1.0
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.6.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.14.0
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing==3.6.4
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.10
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zope.interface==4.3.2
